auth gate

+++ app/Http/Controllers/QuestionsController.php
@@ -65,6 +65,8 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if(\Gate::denies('update-question',$question))
+            abort(403,'Access Denied');
         return view('questions.edit')->with('question',$question);
     }
 
@@ -77,6 +79,9 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(\Gate::denies('update-question',$question))
+            abort(403,'Access Denied');
+
         $question->update($request->only(['title','body']));
         return redirect('questions')->with('success','Your Question Has Been Updated');
     }
@@ -89,6 +94,9 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('delete-question',$question))
+            abort(403,'Access Denied');
+
         $question->delete();
         return redirect("questions")->with('success','Question Deleted Successfully');
     }
